Commit with adding inventory

+++ AccountsWebApi/Controllers/ProfilesController.cs
@@ -29,7 +29,7 @@ namespace AccountsWebApi.Controllers
         // GET: api/Profiles
         [HttpGet]
         [Authorize]
-        public async Task<ActionResult<IEnumerable<Profile>>> Getprofiles(string cid)
+        public async Task<ActionResult<IEnumerable<Profile>>> Getprofiles()
         {
             try
             {
@@ -37,7 +37,7 @@ namespace AccountsWebApi.Controllers
                 var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
                 if (claimresponse.isAuth == true)
                 {
-                    return await _context.profiles.Where(x => x.companyId == cid && x.status == "Active").ToListAsync();
+                    return await _context.profiles.Where(x => x.companyId == claimresponse.companyId && x.status == "Active").ToListAsync();
                 }
                 return Unauthorized();
             }
@@ -51,7 +51,7 @@ namespace AccountsWebApi.Controllers
         // GET: api/Profiles/5
         [HttpGet("getProfile")]
         [Authorize]
-        public async Task<ActionResult<Profile>> GetProfile(string id, int pid)
+        public async Task<ActionResult<Profile>> GetProfile(int pid)
         {
             try
             {
@@ -61,7 +61,7 @@ namespace AccountsWebApi.Controllers
                 {
                     var getRoleId = await _context.profiles
                    .Join(_context.userAndProfiles, d => d.profileAutoId, sd => sd.profileAutoId, (d, sd) => new { d, sd })
-                   .Where(x => x.sd.companyId == id && x.d.companyId == id && x.d.profileAutoId == pid)
+                   .Where(x => x.sd.companyId == claimresponse.companyId && x.d.companyId == claimresponse.companyId && x.d.profileAutoId == pid)
                    .Select(result => new
                    {
                        result.sd.userAutoId,
@@ -95,65 +95,69 @@ namespace AccountsWebApi.Controllers
                 var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
                 if (claimresponse.isAuth == true)
                 {
-                    var listUserName = profile.listUserName;
-                    var listDbUsername = new List<string>();
-                    var getProfileUser = _context.userAndProfiles.Where(x => x.profileAutoId == profile.profileAutoId && profile.companyId == profile.companyId).Select(x => x.userAutoId);
-                    foreach (var x in getProfileUser)
+                    if (ProfileExists(id))
                     {
-                        listDbUsername.Add(x.ToString());
-                        Console.WriteLine("DB Usernames: " + x);
-                    }
-                    foreach (var x in listUserName)
-                    {
-                        Console.WriteLine("New Usernames: " + x);
-
-                    }
-
-                    var resultListLeft = listDbUsername.Except(listUserName).ToList();
-                    var resultListRight = listUserName.Except(listDbUsername).ToList();
-                    var rll = new List<string>();
-                    var rlr = new List<string>();
-                    if (resultListLeft != null)
-                    {
-                        foreach (var x in resultListLeft)
+                        var listUserName = profile.listUserName;
+                        var listDbUsername = new List<string>();
+                        var getProfileUser = _context.userAndProfiles.Where(x => x.profileAutoId == profile.profileAutoId && profile.companyId == claimresponse.companyId).Select(x => x.userAutoId);
+                        foreach (var x in getProfileUser)
                         {
-                            rll.Add(x);
+                            listDbUsername.Add(x.ToString());
+                            Console.WriteLine("DB Usernames: " + x);
                         }
-                    }
-                    if (resultListRight != null)
-                    {
-                        foreach (var x in resultListRight)
+                        foreach (var x in listUserName)
                         {
-                            rlr.Add(x);
+                            Console.WriteLine("New Usernames: " + x);
+
                         }
-                    }
-                    if (rll != null)
-                    {
-                        foreach (var item in rll)
+
+                        var resultListLeft = listDbUsername.Except(listUserName).ToList();
+                        var resultListRight = listUserName.Except(listDbUsername).ToList();
+                        var rll = new List<string>();
+                        var rlr = new List<string>();
+                        if (resultListLeft != null)
                         {
-                            var delUser = _context.userAndProfiles.Where(x => x.companyId == profile.companyId && x.profileAutoId == profile.profileAutoId && x.userAutoId == int.Parse(item)).FirstOrDefault();
-                            if (delUser == null)
+                            foreach (var x in resultListLeft)
                             {
-                                return NotFound();
+                                rll.Add(x);
+                            }
+                        }
+                        if (resultListRight != null)
+                        {
+                            foreach (var x in resultListRight)
+                            {
+                                rlr.Add(x);
+                            }
+                        }
+                        if (rll != null)
+                        {
+                            foreach (var item in rll)
+                            {
+                                var delUser = _context.userAndProfiles.Where(x => x.companyId == claimresponse.companyId && x.profileAutoId == profile.profileAutoId && x.userAutoId == int.Parse(item)).FirstOrDefault();
+                                if (delUser == null)
+                                {
+                                    return NotFound();
+                                }
+
+                                _context.userAndProfiles.Remove(delUser);
+                            }
+                        }
+                        if (rlr != null)
+                        {
+                            foreach (var item in rlr)
+                            {
+                                Userandprofile up = new Userandprofile();
+                                up.profileAutoId = profile.profileAutoId;
+                                up.userAutoId = int.Parse(item);
+                                up.companyId = claimresponse.companyId;
+                                _context.userAndProfiles.Add(up);
                             }
 
-                            _context.userAndProfiles.Remove(delUser);
                         }
+                        await _context.SaveChangesAsync();
+                        return Ok();
                     }
-                    if (rlr != null)
-                    {
-                        foreach (var item in rlr)
-                        {
-                            Userandprofile up = new Userandprofile();
-                            up.profileAutoId = profile.profileAutoId;
-                            up.userAutoId = int.Parse(item);
-                            up.companyId = profile.companyId;
-                            _context.userAndProfiles.Add(up);
-                        }
-
-                    }
-                    await _context.SaveChangesAsync();
-                    return Ok();
+                    return NotFound();
                 }
                 return Unauthorized();
             }
@@ -177,7 +181,7 @@ namespace AccountsWebApi.Controllers
                 if (claimresponse.isAuth == true)
                 {
                     int getprofileautoid = 0;
-                    var compId = _context.profiles.Where(d => d.companyId == profile.companyId).Select(d => d.profileId).ToList();
+                    var compId = _context.profiles.Where(d => d.companyId == claimresponse.companyId).Select(d => d.profileId).ToList();
 
                     var autoid = "";
                     if (compId.Count > 0)
@@ -193,7 +197,7 @@ namespace AccountsWebApi.Controllers
                         p.profileId = comid;
                         p.profileName = profile.profileName;
                         p.status = profile.status;
-                        p.companyId = profile.companyId;
+                        p.companyId = claimresponse.companyId;
                         _context.profiles.Add(p);
                         await _context.SaveChangesAsync();
                         getprofileautoid = p.profileAutoId;
@@ -206,7 +210,7 @@ namespace AccountsWebApi.Controllers
                         p.profileId = comid;
                         p.profileName = profile.profileName;
                         p.status = profile.status;
-                        p.companyId = profile.companyId;
+                        p.companyId = claimresponse.companyId;
                         _context.profiles.Add(p);
                         await _context.SaveChangesAsync();
                         getprofileautoid = p.profileAutoId;
@@ -219,7 +223,7 @@ namespace AccountsWebApi.Controllers
                         Userandprofile up = new Userandprofile();
                         up.userAutoId = int.Parse(items);
                         up.profileAutoId = getprofileautoid;
-                        up.companyId = profile.companyId;
+                        up.companyId = claimresponse.companyId;
                         _context.userAndProfiles.Add(up);
                         await _context.SaveChangesAsync();
                     }
@@ -247,16 +251,20 @@ namespace AccountsWebApi.Controllers
                 var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
                 if (claimresponse.isAuth == true)
                 {
-                    var profile = await _context.profiles.FindAsync(id);
-                    if (profile == null)
+                    if (ProfileExists(id))
                     {
-                        return NotFound();
+                        var profile = await _context.profiles.FindAsync(id);
+                        if (profile == null)
+                        {
+                            return NotFound();
+                        }
+
+                        _context.profiles.Remove(profile);
+                        await _context.SaveChangesAsync();
+
+                        return Ok();
                     }
-
-                    _context.profiles.Remove(profile);
-                    await _context.SaveChangesAsync();
-
-                    return NoContent();
+                    return NotFound();
                 }
                 return Unauthorized();
